Refactor namespaces in DRM UnitTests project
Refactored namespace structure and updated import statements across multiple test files in the `DigitalRightsManagement.UnitTests.Analyzers` project.

- Changed `Analyzers.Analyzers` to `Analyzers.Analysis` in several test files.
- Changed `Analyzers.SourceGenerators` to `Analyzers.Generators` in generator test files.
- Changed `Analyzers.Verifiers` to `Analyzers.Infrastructure.Verification` in verifier test files.
- Changed `Analyzers` to `Analyzers.Infrastructure.TestFramework` in test framework files.
- Updated import statements to reflect these namespace changes.

These changes improve code organization and clarity by grouping related classes and methods into more specific namespaces.
